@@ -3,12 +3,18 @@ module github.com/pathecho
 require github.com/gorilla/mux v1.8.0
 
 require (
+	github.com/golang/protobuf v1.5.2 // indirect
+	golang.org/x/crypto v0.0.0-20190308221718-c2843e01d9a2 // indirect
+	golang.org/x/net v0.5.0 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
+	google.golang.org/protobuf v1.28.0 // indirect
+)
+
+require (
 	github.com/coreos/go-oidc v2.2.1+incompatible
-	github.com/gorilla/context v1.1.1
 	github.com/pquerna/cachecontrol v0.1.0 // indirect
-	github.com/rs/zerolog v1.29.0
 	golang.org/x/oauth2 v0.4.0
 	gopkg.in/square/go-jose.v2 v2.6.0 // indirect
 )
 
-go 1.13
+go 1.19
